Added new versions
Added new versions sent by Ian
